Refactor of if statements

+++ Cleaner.py
@@ -31,25 +31,24 @@ def get_files():
             if(name == "desktop.ini"):
                 continue
             #Only want to assess files that have not been used for at least a week or in the current month
-            if(os.path.exists(name)):
-                if(is_file_one_week_old(name)):
+            if os.path.exists(name) and is_file_one_week_old(name):
                     file_paths.append(os.path.join(path, name))
 
 #Checks if the file can be deleted and deletes it
 def assess_file():
 
     for item in file_paths:
-        if(os.path.exists(item) and os.path.isfile(item)):
-            if(prompt_delete(item)):
+        if os.path.exists(item) and os.path.isfile(item):
+            if prompt_delete(item):
                 delete_file(item)
-                print()
+
 
 def is_file_one_week_old(file):
 
     date_accessed = datetime.datetime.fromtimestamp(os.path.getmtime(file))
 
-    if (datetime.datetime.utcnow().month != date_accessed.month) or \
-                    (datetime.datetime.utcnow().day - date_accessed.day) > 7:
+    if datetime.datetime.utcnow().month != date_accessed.month or \
+                    datetime.datetime.utcnow().day - date_accessed.day > 7:
         return True
     else:
         return False
@@ -63,15 +62,15 @@ def prompt_delete(file):
     while user_answer not in ["0", "1"]:
         user_answer= raw_input("Do you want to delete " + file + "?")
 
-    if(user_answer == "0"):
+    if user_answer == "0":
         return False
-    if(user_answer == "1"):
+    if user_answer == "1":
         return True
 
 
 
 def delete_file(path):
-     if(os.path.isfile(path)):
+     if os.path.isfile(path) and os.path.exists(path):
          os.remove(path)
          print(os.path.basename(path) + "deleted.")
      else:
